Move client run logic to module
