+++ crates/comenq/src/lib.rs
@@ -1,11 +1,11 @@
 //! Library utilities for the `comenq` CLI.
 
 use clap::Parser;
-use comenq_lib::CommentRequest;
 use std::path::PathBuf;
-use thiserror::Error;
-use tokio::io::AsyncWriteExt;
-use tokio::net::UnixStream;
+
+mod client;
+
+pub use client::{ClientError, run};
 
 /// Command line arguments for the `comenq` client.
 #[derive(Debug, Clone, Parser)]
@@ -35,76 +35,11 @@ fn validate_repo_slug(s: &str) -> Result<String, String> {
     }
 }
 
-/// Errors that can occur when interacting with the daemon.
-#[derive(Debug, Error)]
-pub enum ClientError {
-    /// Connecting to the daemon failed.
-    #[error("failed to connect to daemon: {0}")]
-    Connect(#[from] std::io::Error),
-    /// Serialising the request failed.
-    #[error("failed to serialise request: {0}")]
-    Serialise(#[from] serde_json::Error),
-    /// Writing the request to the socket failed.
-    #[error("failed to write to daemon: {0}")]
-    Write(#[source] std::io::Error),
-}
-
-/// Send a `CommentRequest` to the daemon.
-///
-/// # Examples
-///
-/// ```no_run
-/// # use comenq::{Args, run};
-/// # use std::path::PathBuf;
-/// # async fn try_run() -> Result<(), comenq::ClientError> {
-/// let args = Args {
-///     repo_slug: "owner/repo".into(),
-///     pr_number: 1,
-///     comment_body: String::from("Hi"),
-///     socket: PathBuf::from("/run/comenq/socket"),
-/// };
-/// run(args).await?;
-/// # Ok(())
-/// # }
-/// ```
-pub async fn run(args: Args) -> Result<(), ClientError> {
-    let (owner, repo) = parse_slug(&args.repo_slug);
-    let request = CommentRequest {
-        owner,
-        repo,
-        pr_number: args.pr_number,
-        body: args.comment_body,
-    };
-
-    let payload = serde_json::to_vec(&request)?;
-
-    let mut stream = UnixStream::connect(&args.socket)
-        .await
-        .map_err(ClientError::Connect)?;
-    stream
-        .write_all(&payload)
-        .await
-        .map_err(ClientError::Write)?;
-    let _ = stream.shutdown().await;
-    Ok(())
-}
-
-fn parse_slug(slug: &str) -> (String, String) {
-    let mut parts = slug.splitn(2, '/');
-    let owner = parts.next().unwrap_or_default().to_owned();
-    let repo = parts.next().unwrap_or_default().to_owned();
-    (owner, repo)
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{Args, ClientError, run};
+    use super::Args;
     use clap::Parser;
-    use comenq_lib::CommentRequest;
     use rstest::rstest;
-    use tempfile::tempdir;
-    use tokio::io::AsyncReadExt;
-    use tokio::net::UnixListener;
 
     #[rstest]
     #[case("octocat/hello-world", 1, "Hi")]
@@ -125,49 +60,5 @@ mod tests {
     fn rejects_invalid_slug(#[case] slug: &str) {
         let result = Args::try_parse_from(["comenq", slug, "1", "Hi"]);
         assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn run_sends_request() {
-        let dir = tempdir().expect("temp dir");
-        let socket = dir.path().join("sock");
-        let listener = UnixListener::bind(&socket).expect("bind socket");
-
-        let accept = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.expect("accept");
-            let mut buf = Vec::new();
-            stream.read_to_end(&mut buf).await.expect("read");
-            serde_json::from_slice::<CommentRequest>(&buf).expect("deserialize")
-        });
-
-        let args = Args {
-            repo_slug: "octocat/hello-world".into(),
-            pr_number: 1,
-            comment_body: "Hi".into(),
-            socket: socket.clone(),
-        };
-
-        run(args).await.expect("run succeeds");
-        let req = accept.await.expect("join");
-        assert_eq!(req.owner, "octocat");
-        assert_eq!(req.repo, "hello-world");
-        assert_eq!(req.pr_number, 1);
-        assert_eq!(req.body, "Hi");
-    }
-
-    #[tokio::test]
-    async fn run_errors_when_socket_missing() {
-        let dir = tempdir().expect("temp dir");
-        let socket = dir.path().join("nosock");
-
-        let args = Args {
-            repo_slug: "octocat/hello-world".into(),
-            pr_number: 1,
-            comment_body: "Hi".into(),
-            socket: socket.clone(),
-        };
-
-        let err = run(args).await.expect_err("should error");
-        assert!(matches!(err, ClientError::Connect(_)));
     }
 }
