+++ docs/comenq-design.md
@@ -336,11 +336,11 @@ This client is a self-contained, robust utility. It provides clear error
 messages for common failure modes, such as an invalid repository slug or the
 inability to connect to the daemon, guiding the user toward a resolution.
 
-The production code exposes a `run` function in the `comenq` crate. The binary
-parses the CLI arguments and delegates to this function, enabling the test
-suite to exercise the network logic directly. Any failures to serialise the
-request or communicate with the daemon are surfaced via a small `ClientError`
-enumeration.
+The production code exposes a `run` function in the `comenq` crate. This logic
+resides in a dedicated `client` module to keep the argument parser focused. The
+binary parses the CLI arguments and delegates to `run`, allowing the test suite
+to exercise the network code directly. Any failures to serialise the request or
+communicate with the daemon are surfaced via a small `ClientError` enumeration.
 
 ## Section 3: Design of the `comenqd` Daemon
