+++ docs/roadmap.md
@@ -22,11 +22,11 @@
 - [x] Implement the client's `main` function to connect to the daemon's Unix
   Domain Socket using `tokio::net::UnixStream`.
 
-- [x] Serialize the `CommentRequest` payload to JSON and write it to the socket
-  stream.
+- [x] Serialize the `CommentRequest` payload to JSON and write it to the
+  socket stream.
 
-- [x] Implement robust error handling and user feedback for connection failures
-  or serialization errors.
+- [x] Implement robust error handling and user feedback for connection
+  failures or serialization errors.
 
 ## Milestone 3: `comenqd` Daemon Core
