+++ tests/steps/mod.rs
@@ -1,6 +1,6 @@
-pub mod comment_steps;
-pub use comment_steps::CommentWorld;
 pub mod cli_steps;
 pub use cli_steps::CliWorld;
 pub mod client_main_steps;
 pub use client_main_steps::ClientWorld;
+pub mod comment_steps;
+pub use comment_steps::CommentWorld;
